requirements/common.txt
BeautifulSoup==3.2.1
Django==1.4.2
Fabric==1.4.3
South==0.7.6
django-grappelli==2.4.2
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
python-memcached==1.48
pytz==2012f
sorl-thumbnail==11.12
http://google-visualization-python.googlecode.com/files/gviz_api_py-1.8.0.tar.gz#egg=gviz-api.py
https://github.com/smarnach/pyexiftool/tarball/master#egg=PyExifTool
